requirements.txt
altgraph==0.17.5
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
ntplib==0.4.0
packaging==26.0
pefile==2024.8.26
pycparser==3.0
pycryptodome==3.23.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
pywin32-ctypes==0.2.3
setuptools==82.0.0
twofish==0.3.0
